requirements.txt
Pillow
opencv-python
Pillow
numpy

